(tae): small refactor 3
